pyproject.toml
[tool.poetry]
name = "NeRF-extension"
version = "0.1.0"
description = "A mesh reconstruction tool build around Neural Radiance Fields"
authors = ["Cristian Chivrga <cchivriga@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
imageio = "^2.8.0"
numpy = "^1.18.4"
pyyaml = "^5.3.1"
tensorboard = "^2.2.1"
torchvision = "^0.5.0"
tqdm = "^4.46.0"
opencv-python-headless = "^4.2.0"
matplotlib = "^3.2.1"
trimesh = "^3.6.38"
pyrender = "^0.1.40"
pycollada = "^0.7.1"
scipy = "^1.4.1"
scikit-image = "^0.17.2"
torchsearchsorted = {git = "https://github.com/aliutkus/torchsearchsorted" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
